Several things:
(1) Add a bunch of tedious type casts to eliminate unnecessary compiler warnings on Windows X64 builds.

(2) Some EthernetTap work to integrate Windows custom IOCTL for multicast group lookup (not done quite yet).

(3) Dump some more info in selftest to make sure our Windows path lookup functions are returning sane results.

+++ node/C25519.hpp
@@ -73,7 +73,7 @@ public:
 		throw()
 	{
 		Pair kp;
-		Utils::getSecureRandom(kp.priv.data,kp.priv.size());
+		Utils::getSecureRandom(kp.priv.data,(unsigned int)kp.priv.size());
 		_calcPubDH(kp);
 		_calcPubED(kp);
 		return kp;
@@ -98,7 +98,7 @@ public:
 	{
 		Pair kp;
 		void *const priv = (void *)kp.priv.data;
-		Utils::getSecureRandom(priv,kp.priv.size());
+		Utils::getSecureRandom(priv,(unsigned int)kp.priv.size());
 		_calcPubED(kp); // do Ed25519 key -- bytes 32-63 of pub and priv
 		do {
 			++(((uint64_t *)priv)[1]);
